Restructure folder app
