+++ app/main/browser-window.js
