+++ app/main/note-tray.js
