+++ app/media-capturer/dragsnip/capture-editor.js
