+++ app/media-capturer/dragsnip/capture-main.js
