+++ app/media-capturer/dragsnip/capture-renderer.js
