+++ app/media-capturer/dragsnip/desktop-capturer.js
