+++ app/media-capturer/dragsnip/utils.js
