+++ app/renderer/note-manager.js
